# api/docs/swagger_v3.json
{"components":{"schemas":{"main.CreatePostPayload":{"properties":{"content":{"maxLength":1000,"type":"string"},"tags":{"items":{"type":"string"},"type":"array"},"title":{"maxLength":100,"type":"string"}},"required":["content","title"],"type":"object"},"main.CreateUserTokenPayload":{"properties":{"email":{"maxLength":255,"type":"string"},"password":{"maxLength":72,"minLength":3,"type":"string"}},"required":["email","password"],"type":"object"},"main.RegisterUserPayload":{"properties":{"email":{"maxLength":255,"type":"string"},"password":{"maxLength":72,"minLength":3,"type":"string"},"username":{"maxLength":100,"type":"string"}},"required":["email","password","username"],"type":"object"},"main.UpdatePostPayload":{"properties":{"content":{"maxLength":1000,"type":"string"},"title":{"maxLength":100,"type":"string"}},"type":"object"},"main.UserWithToken":{"properties":{"created_at":{"type":"string"},"email":{"type":"string"},"id":{"type":"integer"},"is_active":{"type":"boolean"},"role":{"$ref":"#/components/schemas/store.Role"},"role_id":{"type":"integer"},"token":{"type":"string"},"username":{"type":"string"}},"required":["created_at","email","id","is_active","role","role_id","username"],"type":"object"},"store.Comment":{"properties":{"content":{"type":"string"},"created_at":{"type":"string"},"id":{"type":"integer"},"post_id":{"type":"integer"},"user":{"$ref":"#/components/schemas/store.User"},"user_id":{"type":"integer"}},"type":"object"},"store.Post":{"properties":{"comments":{"items":{"$ref":"#/components/schemas/store.Comment"},"type":"array"},"content":{"type":"string"},"created_at":{"type":"string"},"id":{"type":"integer"},"tags":{"items":{"type":"string"},"type":"array"},"title":{"type":"string"},"updated_at":{"type":"string"},"user":{"$ref":"#/components/schemas/store.User"},"user_id":{"type":"integer"},"version":{"type":"integer"}},"type":"object"},"store.PostWithMetadata":{"properties":{"comments":{"items":{"$ref":"#/components/schemas/store.Comment"},"type":"array"},"comments_count":{"type":"integer"},"content":{"type":"string"},"created_at":{"type":"string"},"id":{"type":"integer"},"tags":{"items":{"type":"string"},"type":"array"},"title":{"type":"string"},"updated_at":{"type":"string"},"user":{"$ref":"#/components/schemas/store.User"},"user_id":{"type":"integer"},"version":{"type":"integer"}},"type":"object"},"store.Role":{"properties":{"description":{"type":"string"},"id":{"type":"integer"},"level":{"type":"integer"},"name":{"type":"string"}},"type":"object"},"store.User":{"properties":{"created_at":{"type":"string"},"email":{"type":"string"},"id":{"type":"integer"},"is_active":{"type":"boolean"},"role":{"$ref":"#/components/schemas/store.Role"},"role_id":{"type":"integer"},"username":{"type":"string"}},"required":["created_at","email","id","is_active","role","role_id","username"],"type":"object"}},"securitySchemes":{"ApiKeyAuth":{"in":"header","name":"Authorization","type":"apiKey"}}},"info":{"contact":{"email":"support@swagger.io","name":"API Support","url":"http://www.swagger.io/support"},"description":"API for GopherSocial, a social network for gohpers","license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0.html"},"termsOfService":"http://swagger.io/terms/","title":"GopherSocial API","version":"1.0.0"},"openapi":"3.0.3","paths":{"/authentication/token":{"post":{"description":"Creates a token for a user","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/main.CreateUserTokenPayload"}}},"description":"User credentials","required":true,"x-originalParamName":"payload"},"responses":{"200":{"content":{"application/json":{"schema":{"type":"string"}}},"description":"Token"},"400":{"content":{"application/json":{"schema":{}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{}}},"description":"Unauthorized"},"500":{"content":{"application/json":{"schema":{}}},"description":"Internal Server Error"}},"summary":"Creates a token","tags":["authentication"]}},"/authentication/user":{"post":{"description":"Registers a user","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/main.RegisterUserPayload"}}},"description":"User credentials","required":true,"x-originalParamName":"payload"},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/main.UserWithToken"}}},"description":"User registered"},"400":{"content":{"application/json":{"schema":{}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{}}},"description":"Internal Server Error"}},"summary":"Registers a user","tags":["authentication"]}},"/health":{"get":{"description":"Healthcheck endpoint","responses":{"200":{"content":{"application/json":{"schema":{"type":"string"}}},"description":"ok"}},"summary":"Healthcheck","tags":["ops"]}},"/posts":{"post":{"description":"Creates a post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/main.CreatePostPayload"}}},"description":"Post payload","required":true,"x-originalParamName":"payload"},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/store.Post"}}},"description":"Created"},"400":{"content":{"application/json":{"schema":{}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{}}},"description":"Unauthorized"},"500":{"content":{"application/json":{"schema":{}}},"description":"Internal Server Error"}},"security":[{"ApiKeyAuth":[]}],"summary":"Creates a post","tags":["posts"]}},"/posts/{id}":{"delete":{"description":"Delete a post by ID","parameters":[{"description":"Post ID","in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"204":{"content":{"application/json":{"schema":{"type":"string"}}},"description":"No Content"},"404":{"content":{"application/json":{"schema":{}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{}}},"description":"Internal Server Error"}},"security":[{"ApiKeyAuth":[]}],"summary":"Deletes a post","tags":["posts"]},"get":{"description":"Fetches a post by ID","parameters":[{"description":"Post ID","in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/store.Post"}}},"description":"OK"},"404":{"content":{"application/json":{"schema":{}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{}}},"description":"Internal Server Error"}},"security":[{"ApiKeyAuth":[]}],"summary":"Fetches a post","tags":["posts"]},"patch":{"description":"Updates a post by ID","parameters":[{"description":"Post ID","in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/main.UpdatePostPayload"}}},"description":"Post payload","required":true,"x-originalParamName":"payload"},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/store.Post"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{}}},"description":"Internal Server Error"}},"security":[{"ApiKeyAuth":[]}],"summary":"Updates a post","tags":["posts"]}},"/users/activate/{token}":{"put":{"description":"Activates/Register a user by invitation token","parameters":[{"description":"Invitation token","in":"path","name":"token","required":true,"schema":{"type":"string"}}],"responses":{"204":{"content":{"application/json":{"schema":{"type":"string"}}},"description":"User activated"},"404":{"content":{"application/json":{"schema":{}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{}}},"description":"Internal Server Error"}},"security":[{"ApiKeyAuth":[]}],"summary":"Activates/Register a user","tags":["users"]}},"/users/feed":{"get":{"description":"Fetches the user feed","parameters":[{"description":"Since","in":"query","name":"since","schema":{"type":"string"}},{"description":"Until","in":"query","name":"until","schema":{"type":"string"}},{"description":"Limit","in":"query","name":"limit","schema":{"type":"integer"}},{"description":"Offset","in":"query","name":"offset","schema":{"type":"integer"}},{"description":"Sort","in":"query","name":"sort","schema":{"type":"string"}},{"description":"Tags","in":"query","name":"tags","schema":{"type":"string"}},{"description":"Search","in":"query","name":"search","schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/store.PostWithMetadata"},"type":"array"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{}}},"description":"Internal Server Error"}},"summary":"Fetches the user feed","tags":["feed"]}},"/users/{id}":{"get":{"description":"Fetches a user profile by ID","parameters":[{"description":"User ID","in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/store.User"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{}}},"description":"Bad Request"},"404":{"content":{"application/json":{"schema":{}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{}}},"description":"Internal Server Error"}},"security":[{"ApiKeyAuth":[]}],"summary":"Fetches a user profile","tags":["users"]}},"/users/{userID}/follow":{"put":{"description":"Follows a user by ID","parameters":[{"description":"User ID","in":"path","name":"userID","required":true,"schema":{"type":"integer"}}],"responses":{"204":{"content":{"application/json":{"schema":{"type":"string"}}},"description":"User followed"},"400":{"content":{"application/json":{"schema":{}}},"description":"User payload missing"},"404":{"content":{"application/json":{"schema":{}}},"description":"User not found"}},"security":[{"ApiKeyAuth":[]}],"summary":"Follows a user","tags":["users"]}},"/users/{userID}/unfollow":{"put":{"description":"Unfollow a user by ID","parameters":[{"description":"User ID","in":"path","name":"userID","required":true,"schema":{"type":"integer"}}],"responses":{"204":{"content":{"application/json":{"schema":{"type":"string"}}},"description":"User unfollowed"},"400":{"content":{"application/json":{"schema":{}}},"description":"User payload missing"},"404":{"content":{"application/json":{"schema":{}}},"description":"User not found"}},"security":[{"ApiKeyAuth":[]}],"summary":"Unfollow a user","tags":["users"]}}}}